+++ src/layout/content-main.component.ts
@@ -1,17 +1,30 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component, Directive } from '@angular/core';
 
 /**
- * @name thy-content-main
+ * 内容主区域布局指令
+ * @name thyContentMain
  * @order 45
+ */
+@Directive({
+    selector: '[thyContentMain]',
+    host: {
+        class: 'thy-layout-content-main'
+    },
+    standalone: true
+})
+export class ThyContentMainDirective {}
+
+/**
+ * 内容主区域布局组件
+ * @name thy-content-main
+ * @order 46
  */
 @Component({
     selector: 'thy-content-main',
     preserveWhitespaces: false,
     template: ` <ng-content></ng-content> `,
     changeDetection: ChangeDetectionStrategy.OnPush,
-    host: {
-        class: 'thy-layout-content-main'
-    },
+    hostDirectives: [ThyContentMainDirective],
     standalone: true
 })
 export class ThyContentMainComponent {}