+++ src/layout/sidebar-footer.component.ts
@@ -1,20 +1,29 @@
-import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, Directive, OnInit } from '@angular/core';
 
 /**
- * @name thy-sidebar-footer
+ * 侧边栏底部布局指令
+ * @name thySidebarFooter
  * @order 35
  */
-@Component({
-    selector: 'thy-sidebar-footer',
-    template: ` <ng-content></ng-content> `,
-    changeDetection: ChangeDetectionStrategy.OnPush,
+@Directive({
+    selector: '[thySidebarFooter]',
     host: {
         class: 'sidebar-footer'
     },
     standalone: true
 })
-export class ThySidebarFooterComponent implements OnInit {
-    constructor() {}
+export class ThySidebarFooterDirective {}
 
-    ngOnInit(): void {}
-}
+/**
+ * 侧边栏底部布局组件
+ * @name thy-sidebar-footer
+ * @order 36
+ */
+@Component({
+    selector: 'thy-sidebar-footer',
+    template: ` <ng-content></ng-content> `,
+    changeDetection: ChangeDetectionStrategy.OnPush,
+    hostDirectives: [ThySidebarFooterDirective],
+    standalone: true
+})
+export class ThySidebarFooterComponent {}