+++ src/select/custom-select.spec.ts
@@ -20,10 +20,16 @@ import { ThyFormModule } from '../form';
 import { ThyOptionModule } from '../shared/option/module';
 import { ThyOptionComponent } from '../shared/option/option.component';
 import { DOWN_ARROW, END, ENTER, ESCAPE, HOME } from '../util/keycodes';
-import { SelectMode, THY_SELECT_PANEL_MIN_WIDTH, ThySelectCustomComponent } from './custom-select/custom-select.component';
+import {
+    SelectMode,
+    THY_SELECT_PANEL_MIN_WIDTH,
+    ThySelectCustomComponent,
+    ThySelectOptionModel
+} from './custom-select/custom-select.component';
 import { ThySelectModule } from './module';
 import { THY_SELECT_CONFIG, THY_SELECT_SCROLL_STRATEGY, ThyDropdownWidthMode } from './select.config';
 import { POSITION_MAP, ThyPlacement } from 'ngx-tethys/core';
+import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 
 @Component({
     selector: 'thy-select-basic-test',
@@ -677,6 +683,23 @@ class SelectDropdownWidthComponent {
     selectedValue = this.options[0].value;
 }
 
+@Component({
+    selector: 'thy-select-width-thy-options',
+    template: `
+        <div style="width:100px">
+            <thy-custom-select [thyOptions]="options" class="select1" [(ngModel)]="selectedValue"> </thy-custom-select>
+        </div>
+    `
+})
+class SelectWidthThyOptionsComponent {
+    options: ThySelectOptionModel[] = [
+        { value: 'apple', label: '苹果' },
+        { value: 'orange', label: '橘子' },
+        { value: 'banana', label: '香蕉' }
+    ];
+    selectedValue = this.options[0].value;
+}
+
 describe('ThyCustomSelect', () => {
     let overlayContainer: OverlayContainer;
     let overlayContainerElement: HTMLElement;
@@ -684,7 +707,7 @@ describe('ThyCustomSelect', () => {
 
     function configureThyCustomSelectTestingModule(declarations: any[], providers: any[] = []) {
         TestBed.configureTestingModule({
-            imports: [ThyFormModule, ThyOptionModule, ThySelectModule, ReactiveFormsModule, FormsModule],
+            imports: [ThyFormModule, ThyOptionModule, ThySelectModule, ReactiveFormsModule, FormsModule, NoopAnimationsModule],
             declarations: declarations,
             providers: [bypassSanitizeProvider, ...providers]
         }).compileComponents();
@@ -1602,7 +1625,7 @@ describe('ThyCustomSelect', () => {
             trigger.click();
             fixture.detectChanges();
 
-            const groups = fixture.componentInstance.select.optionGroups.toArray();
+            const groups = fixture.componentInstance.select.contentGroups.toArray();
             const options = fixture.componentInstance.select.options.toArray();
             const input = fixture.debugElement.query(By.css('.search-input-field')).nativeElement;
             typeInElement('Cat', input);
@@ -1649,7 +1672,7 @@ describe('ThyCustomSelect', () => {
 
             expect(overlayContainerElement.querySelector('thy-empty')).not.toBeTruthy();
 
-            const groups = fixture.componentInstance.select.optionGroups.toArray();
+            const groups = fixture.componentInstance.select.contentGroups.toArray();
             const input = fixture.debugElement.query(By.css('.search-input-field')).nativeElement;
             typeInElement('cat2', input);
 
@@ -1679,7 +1702,7 @@ describe('ThyCustomSelect', () => {
             fixture.componentInstance.disabled = true;
             fixture.detectChanges();
             const removeIcon2 = fixture.debugElement.query(By.css('.choice-remove-link'));
-            const choice = fixture.debugElement.query(By.css('.choice')).nativeElement as HTMLElement;
+            const choice = fixture.debugElement.query(By.css('.choice-item')).nativeElement as HTMLElement;
             tick();
             expect(choice.classList.contains('disabled')).toBeTruthy();
             expect(removeIcon2).toBeNull();
@@ -2480,6 +2503,59 @@ describe('ThyCustomSelect', () => {
             tick(2000);
 
             expect(componentFocusSpy).toHaveBeenCalled();
+        }));
+    });
+
+    describe('use thyOptions', () => {
+        beforeEach(() => {
+            configureThyCustomSelectTestingModule([SelectWidthThyOptionsComponent]);
+        });
+
+        it('should has correct thyOption component', fakeAsync(() => {
+            const fixture = TestBed.createComponent(SelectWidthThyOptionsComponent);
+            fixture.detectChanges();
+            const trigger = fixture.debugElement.query(By.css('.form-control-custom')).nativeElement;
+            trigger.click();
+            fixture.detectChanges();
+            flush();
+            fixture.detectChanges();
+            expect(overlayContainerElement.textContent).toContain('香蕉');
+            expect(overlayContainerElement.textContent).toContain('苹果');
+            expect(overlayContainerElement.textContent).toContain('橘子');
+        }));
+
+        it('should get correct value when click option', fakeAsync(() => {
+            const fixture = TestBed.createComponent(SelectWidthThyOptionsComponent);
+            fixture.detectChanges();
+            const trigger = fixture.debugElement.query(By.css('.form-control-custom')).nativeElement;
+            trigger.click();
+            fixture.detectChanges();
+            flush();
+            fixture.detectChanges();
+            const option = overlayContainerElement.querySelector('thy-option') as HTMLElement;
+            option.click();
+            fixture.detectChanges();
+            flush();
+            expect(fixture.componentInstance.selectedValue).toEqual(fixture.componentInstance.options[0].value);
+        }));
+
+        it('should show correct when option has group label', fakeAsync(() => {
+            const fixture = TestBed.createComponent(SelectWidthThyOptionsComponent);
+            fixture.detectChanges();
+            fixture.componentInstance.options = [
+                { label: '猫', value: 'cat', groupLabel: 'pet' },
+                { label: '狗', value: 'dog', groupLabel: 'pet' },
+                { label: '猪', value: 'pig' }
+            ];
+            fixture.detectChanges();
+            const trigger = fixture.debugElement.query(By.css('.form-control-custom')).nativeElement;
+            trigger.click();
+            fixture.detectChanges();
+            flush();
+            fixture.detectChanges();
+            const optionGroup = overlayContainerElement.querySelector('thy-option-group') as HTMLElement;
+            const groupName = optionGroup.querySelector('.group-name') as HTMLElement;
+            expect(groupName.innerText).toEqual(fixture.componentInstance.options[0].groupLabel);
         }));
     });
 });