+++ src/shared/select/select-control/select-control.component.ts
@@ -21,6 +21,7 @@ import { FormsModule } from '@angular/forms';
 import { ThyIconComponent } from 'ngx-tethys/icon';
 import { ThyTagComponent } from 'ngx-tethys/tag';
 import { SelectOptionBase } from '../../option/select-option-base';
+import { ThyGridModule } from 'ngx-tethys/grid';
 
 export type SelectControlSize = 'sm' | 'md' | 'lg' | '';
 
@@ -32,7 +33,7 @@ export type SelectControlSize = 'sm' | 'md' | 'lg' | '';
     templateUrl: './select-control.component.html',
     changeDetection: ChangeDetectionStrategy.OnPush,
     standalone: true,
-    imports: [FormsModule, NgClass, NgIf, NgStyle, NgFor, ThyTagComponent, NgTemplateOutlet, ThyIconComponent],
+    imports: [FormsModule, NgClass, NgIf, NgStyle, NgFor, ThyTagComponent, NgTemplateOutlet, ThyIconComponent, ThyGridModule],
     host: {
         '[class.select-control-borderless]': 'thyBorderless'
     }
@@ -179,6 +180,8 @@ export class ThySelectControlComponent implements OnInit {
 
     @Input() @InputBoolean() thyBorderless = false;
 
+    @Input() thyPreset: string = '';
+
     @Output()
     thyOnSearch = new EventEmitter<string>();
 
@@ -205,7 +208,7 @@ export class ThySelectControlComponent implements OnInit {
         } else {
             showSelectedValue = true;
         }
-        return { display: showSelectedValue ? 'block' : 'none' };
+        return { display: showSelectedValue ? 'flex' : 'none' };
     }
 
     get placeholderStyle(): { [key: string]: string } {