+++ src/switch/examples/loading/loading.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'thy-switch-loading-example',
+    templateUrl: './loading.component.html'
+})
+export class ThySwitchLoadingExampleComponent {
+    isChecked: Boolean = true;
+    noChecked: Boolean = false;
+}